requirements.txt
customtkinter==5.2.1
ctktooltip==0.8
pillow==10.1.0
packaging==23.2